Address file-node review findings
Reject empty interior file-node path segments as invalid input and keep the
path-preparation error shape aligned with `FileNodeLookupError`.

Split the database test module so file-node scenario bodies and the embedded
PostgreSQL harness live in focused child modules. Ensure the embedded server is
stopped after migration or connection failures, and propagate category fixture
setup errors instead of panicking inside the helper.

Resolve Cargo through an absolute path for Makefile lint gates and surface
PostgreSQL cleanup failures from `PostgresTestDb::drop`.
